// src/features/reports/report.types.ts
export type ReportCashieredRepairOrder = {
  id: string;

  roNumber: number;

  customerName: string;

  vehicle: string;

  cashieredAt: string;

  cashierName: string | null;

  invoiceTotal: number;

  cashierStatus: string;
};

//************************************************************** */

export type ReportPickedUpRepairOrder = {
  id: string;

  roNumber: number;

  customerName: string;

  vehicle: string;

  pickedUpAt: string;

  releasedBy: string | null;

  pickupRecipient: string | null;

  cashierStatus: string;

  pickupStatus: string;
};